% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_to_num.R
\name{cat_to_num}
\alias{cat_to_num}
\title{Change Categorical Values to Numeric}
\usage{
cat_to_num(
  data_participants = read_participants(),
  share_reduced = 0.7,
  adjustment_less = 0.8,
  adjustment_more = 1.2
)
}
\arguments{
\item{data_participants}{A dataframe containing the columns \code{share}, \code{adjustment}, and \code{age}.}

\item{share_reduced}{A numeric value to assign when \code{share} is "reduced". Default is \code{0.7}.}

\item{adjustment_less}{A numeric value to assign when \code{adjustment} is "less". Default is \code{0.8}.}

\item{adjustment_more}{A numeric value to assign when \code{adjustment} is "more". Default is \code{1.2}.}
}
\value{
A dataframe with transformed values.
}
\description{
This function converts categorical descriptive values in the columns \code{share}, \code{adjustment}, and \code{age} to numeric values.
}
\examples{
data_participants <- cat_to_num()
}
