% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_costs.R
\name{check_costs}
\alias{check_costs}
\title{Read and check cost data}
\usage{
check_costs(
  cost_df = read.csv(system.file("costs.csv", package = "costsplitter"))
)
}
\arguments{
\item{cost_df}{A character string specifying data. Defaults to the read.csv(system.file("costs.csv", package = "costsplitter")), which is the file included in the costsplitter package.}
}
\value{
A tibble containing the cost data read from the CSV file.
}
\description{
This function reads cost data from a dataset and checks if it works for the pourpose of this package.
}
\examples{
data_costs <- costsplitter::check_costs()
}
