% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_costsplitter_data.R
\name{check_costsplitter_data}
\alias{check_costsplitter_data}
\title{Check the dataset for conformity with costsplitter prerequisites}
\usage{
check_costsplitter_data(
  data = readr::read_csv(system.file("participants.csv", package = "costsplitter"),
    show_col_types = FALSE),
  var_pay = "pay_",
  var_share = "share_",
  value_full = 1,
  value_reduced = 0.8,
  value_half = 0.5,
  value_some = 0.3
)
}
\arguments{
\item{data}{A data frame containing the cost-sharing data.}

\item{var_pay}{A string indicating the prefix for payment columns (default is "pay_").}

\item{var_share}{A string indicating the prefix for share columns (default is "share_").}

\item{value_full}{A numeric value representing the full share (default is 1).}

\item{value_reduced}{A numeric value representing the reduced share (default is 0.8).}

\item{value_half}{A numeric value representing the half share (default is 0.5).}

\item{value_some}{A numeric value representing the partial share (default is 0.3).}
}
\value{
A message indicating whether the dataset conforms to \code{costsplitter} prerequisites.
Stops execution if the dataset does not conform.
}
\description{
This function checks whether a given dataset conforms to the expected format and requirements
of the \code{costsplitter} package. It ensures the presence and validity of required columns such
as \code{name}, \code{pay_}, \code{share_}, \code{group}, \code{age}, and \code{adjustment}.
}
\examples{
library(dplyr)
library(stringr)
data <- data.frame(
  name = c("Alice", "Bob", "Charlie"),
  pay_breakfast = c(20, 0, 10),
  share_breakfast = c(1, "half", "full"),
  age = c(30, "kid", NA),
  adjustment = c("more", NA, "less")
)

check_costsplitter_data(data)

}
