% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{divide_costs}
\alias{divide_costs}
\title{Divide Costs Based on Weighted Factors}
\usage{
divide_costs(df = change_categorical_to_numeric(read_data()), cost = 5500)
}
\arguments{
\item{df}{A dataframe containing the columns \code{agegroup}, \code{age}, \code{adjustment}, and \code{present}. Default is \code{df}.}

\item{cost}{The total cost to be divided among individuals. Default is \code{5500}.}
}
\value{
A dataframe with an additional column \code{to_pay}, representing the calculated cost each individual needs to pay based on their weight.
}
\description{
This function calculates how much each individual should pay, based on various weighted factors such as age group, adjustment, and presence.
}
\details{
The function computes a \code{weight} for each individual based on:
\itemize{
\item Initializing \code{weight} to \code{1}.
\item If \code{agegroup} is \code{"kid"}, the weight is updated to \code{weight * 0.0555 * age}.
\item The weight is further multiplied by values in \code{adjustment} and \code{present} columns.
\item The total cost (\code{cost}) is divided based on the sum of weights to determine each individual's share.
}
}
\examples{
df <- divide_costs()
}
