% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_check_names.R
\name{helper_check_names}
\alias{helper_check_names}
\title{Checks if the correct names are present}
\usage{
helper_check_names(df)
}
\arguments{
\item{df}{A data frame containing an the right columns.}
}
\value{
A data frame with all the right columns
}
\description{
The \code{helper_check_names} function checks if the names are correct. For those that can be missing the variable is created with the correct default value.
}
\examples{
df <- helper_check_names(df =
     tibble::tibble(name = c("Alice", "Alice", "Charlie"),
                    share_tour = c(0.4, 0.6, 0.5),
                    pay_tour = c(100,200,300)))


}
\seealso{
Other helper: 
\code{\link{helper_minimize_payments}()},
\code{\link{helper_process_adjustment}()},
\code{\link{helper_process_age}()},
\code{\link{helper_process_name}()},
\code{\link{helper_process_pay}()},
\code{\link{helper_process_share}()}
}
\concept{helper}
