% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_process_pay.R
\name{helper_process_pay}
\alias{helper_process_pay}
\title{Process Share Data in a Data Frame}
\usage{
helper_process_pay(df)
}
\arguments{
\item{df}{A data frame containing columns that start with "share" and a \code{name} column. The share columns can contain numeric values (0-1) or categorical values ("full", "reduced", "half", "some").}
}
\value{
A tibble with three columns: \code{name}, \code{activity}, and \code{share}. The share values are transformed based on the rules provided, and the data is pivoted to a long format.
}
\description{
The \code{helper_process_share} function processes the \code{share} columns in a given data frame. It pivots the share columns into a long format, validates the data, and converts categorical share values into numeric equivalents. The function also replaces missing or empty values with 0.
}
\details{
The function performs the following steps:
\enumerate{
\item \strong{Pivoting}: Converts the wide format share columns into a long format with \code{activity} and \code{share} columns.
\item \strong{Assertive Tests}: Validates that the \code{share} column contains only numeric values (0-1) or the specified categorical values ("full", "reduced", "half", "some").
\item \strong{Conversion}:
\itemize{
\item Converts categorical values ("full", "reduced", "half", "some") into numeric equivalents (1, 0.7, 0.5, 0.3).
\item Replaces \code{NA} and empty values with 0.
}
\item \strong{Transformation}: Ensures that the \code{share} column is numeric and returns the modified data frame as a tibble.
}
}
\examples{
data("valid_trip_data", package = "costsplitter")
df <- valid_trip_data
data_pay <- helper_process_pay(df)


}
\seealso{
Other helper: 
\code{\link{helper_check_names}()},
\code{\link{helper_minimize_payments}()},
\code{\link{helper_process_adjustment}()},
\code{\link{helper_process_age}()},
\code{\link{helper_process_name}()},
\code{\link{helper_process_share}()}
}
\concept{helper}
