% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{incorrect_share_values}
\alias{incorrect_share_values}
\title{Example Data: Incorrect Share Values}
\format{
A data frame with 4 rows and 8 variables.
}
\usage{
incorrect_share_values
}
\description{
This dataset contains invalid values in the \code{share_} columns, designed to test the package's error handling for incorrect share values.
}
\keyword{internal}
