% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimize_payments.R
\name{minimize_payments}
\alias{minimize_payments}
\title{Minimize Payments Between persons}
\usage{
minimize_payments(df)
}
\arguments{
\item{df}{A tibble or data frame with two columns:
\describe{
\item{person}{A character vector representing the person IDs (e.g., names).}
\item{to_pay}{A numeric vector representing the amount each person needs to pay (positive values) or receive (negative values).}
}}
}
\value{
A tibble with three columns:
\describe{
\item{payer}{The person that is making a payment.}
\item{receiver}{The person that is receiving the payment.}
\item{amount}{The amount of money being transferred.}
}
}
\description{
This function finds the minimum number of transactions required to settle balances between persons where some persons owe money and others need to be paid. It first settles exact matches where a payer owes the exact amount that a receiver needs, and then proceeds to match payers with receivers in a way that minimizes the number of payments.
}
\details{
The function separates the input data frame into payers (persons with positive balances) and receivers (persons with negative balances). It first identifies and settles any exact matches between payers and receivers. Then, for the remaining balances, it matches payers with receivers in a way that minimizes the number of payments.
}
\examples{

df <- dplyr::tibble(
  element = c("Brown", "Smith", "Fisher", "Taylor", "Davis", "Johnson", "Wilson", "ORally"),
  to_pay = c(-995, -125, -80,  210, 245, 300, 365, 80)
)

minimize_payments(df)

}
