% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{missing_columns_data}
\alias{missing_columns_data}
\title{Example Data: Missing Columns Data}
\format{
A data frame with 5 rows and 7 variables.
}
\usage{
missing_columns_data
}
\description{
This dataset is missing the mandatory \code{name} column, designed to test how the package handles missing essential columns.
}
\keyword{internal}
