% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{participants}
\alias{participants}
\title{Example Data: Participants}
\format{
A data frame with 7 rows and 13 variables:
\describe{
\item{id}{Unique identifier for each participant.}
\item{name}{Participant's name, unique.}
\item{group}{Group identifier for participants.}
\item{age}{Age of the participant; can be numeric or categories like 'adult' or 'kid'.}
\item{share_dinnerday1}{Proportion of the dinner cost on day 1 each participant is responsible for.}
\item{pay_dinnerday1}{Amount paid by the participant for dinner on day 1.}
\item{share_day1}{Proportion of the cost on day 1 each participant is responsible for.}
\item{pay_day1}{Amount paid by the participant on day 1.}
\item{share_day2}{Proportion of the cost on day 2 each participant is responsible for.}
\item{pay_day2}{Amount paid by the participant on day 2.}
\item{share_day3}{Proportion of the cost on day 3 each participant is responsible for.}
\item{pay_day3}{Amount paid by the participant on day 3.}
\item{adjustment}{Adjustments to be applied to each participant's cost.}
}
}
\usage{
participants
}
\description{
This dataset contains complete and valid information for testing the \code{costsplitter} package.
It follows all requirements, including proper columns for names, group, age, payments, shares, and adjustments.
}
\keyword{internal}
