% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_costs.R
\name{read_costs}
\alias{read_costs}
\title{Read Cost Data from CSV File}
\usage{
read_costs(path = system.file("costs.csv", package = "dividefair"))
}
\arguments{
\item{path}{A character string specifying the path to the CSV file.
Defaults to the "costs.csv" file included in the dividefair package.}
}
\value{
A tibble containing the cost data read from the CSV file.
}
\description{
This function reads cost data from a CSV file, defaulting to the "costs.csv"
file included in the dividefair package.
}
\examples{
data_costs <- dividefair::read_costs()
}
