% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read data for Grouptraveller}
\usage{
read_data(
  path = system.file("participants.csv", package = "Grouptraveller"),
  var_agegroup = "agegroup",
  value_adults = "adult",
  value_kids = "kid",
  indicator_costs = "cost",
  var_present = "present"
)
}
\arguments{
\item{path}{A file path (string) to an .csv file}
}
\value{
a data frame
}
\description{
Read data for Grouptraveller
}
\examples{
df <- read_data(path = system.file("participants.csv", package = "Grouptraveller"))

}
