% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costsplitter-package.R
\docType{package}
\name{costsplitter-package}
\alias{costsplitter-package}
\title{costsplitter: Divide Costs Fairly Amongst Friends}
\description{
Do the accounting of a group trip.
}
\author{
\strong{Maintainer}: Jakob Schumacher \email{jakob.schumacher@web.de} (\href{https://orcid.org/0000-0002-9894-6085}{ORCID})

}
\keyword{internal}
