% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costsplitter.R
\name{costsplitter}
\alias{costsplitter}
\title{Divide Costs Fairly Among Participants}
\usage{
costsplitter(
  data = readr::read_csv(system.file("participants.csv", package = "costsplitter"),
    show_col_types = FALSE),
  pay_by = "individual"
)
}
\arguments{
\item{data}{The dataset}

\item{pay_by}{A string specifying whether to divide costs by "group" or "individual". Defaults to "group".}
}
\value{
A data frame containing the final payments for each person or group after minimizing the payments.
}
\description{
This function calculates the fair division of costs among participants based on their weights.
It reads the participants' and costs' data from files, calculates the total costs for each participant or group,
and returns the amounts to be paid after applying any payments that have already been made.
}
\examples{
result <- costsplitter(pay_by = "group")
}
