% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{extra_unexpected_columns}
\alias{extra_unexpected_columns}
\title{Example Data: Extra Unexpected Columns}
\format{
A data frame with 4 rows and 10 variables.
}
\usage{
extra_unexpected_columns
}
\description{
This dataset includes extra columns that are not recognized by the package, testing if it can ignore unnecessary data without affecting calculations.
}
\keyword{datasets}
