% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{get_stats}
\alias{get_stats}
\title{Get Statistics on Participants}
\usage{
get_stats(data_participants = read_participants())
}
\arguments{
\item{data_participants}{A data frame containing at least the following columns:
\itemize{
\item \code{agegroup}: A character vector indicating whether the participant is an "adult" or a "kid".
\item \code{share}: A character vector indicating presence with values "x" for stays and "n" for absences.
}}
}
\value{
A list containing:
\itemize{
\item \code{adults}: The total number of adults in the data frame.
\item \code{children}: The total number of children in the data frame.
}
}
\description{
This function calculates statistics regarding the number of adults, children, stays, and absences
in a given data frame based on the \code{agegroup} and \code{share} columns.
}
\examples{
get_stats(read_participants())

}
