% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_process_adjustment.R
\name{helper_process_adjustment}
\alias{helper_process_adjustment}
\title{Process Adjustment Data in a Data Frame}
\usage{
helper_process_adjustment(df = helper_check_names(df))
}
\arguments{
\item{df}{A data frame containing an \code{adjustment} column. The \code{adjustment} column can contain numeric values (0-100) or categorical values ("more" or "less").}
}
\value{
A data frame with the \code{adjustment} column transformed based on the rules provided.
}
\description{
The \code{helper_process_adjustment} function processes the \code{adjustment} column in a given data frame. It validates the data, converts categorical values ("more" and "less") into numeric equivalents, and handles missing values appropriately.
}
\details{
The function performs the following steps:
\enumerate{
\item \strong{Assertive Tests}: Validates that the \code{adjustment} column exists and that it contains only numeric values (0-100) or the categorical values "more" and "less".
\item \strong{Conversion}:
\itemize{
\item Converts "more" to 1.2 and "less" to 0.8.
\item Replaces \code{NA} and empty values with 1.
}
}
}
\examples{
data("valid_trip_data", package = "costsplitter")
data_adjustment <- helper_process_adjustment(df = valid_trip_data)

}
\seealso{
Other helper: 
\code{\link{helper_check_names}()},
\code{\link{helper_minimize_payments}()},
\code{\link{helper_process_age}()},
\code{\link{helper_process_name}()},
\code{\link{helper_process_pay}()},
\code{\link{helper_process_share}()}
}
\concept{helper}
