% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_process_age.R
\name{helper_process_age}
\alias{helper_process_age}
\title{Process Age Data in a Data Frame}
\usage{
helper_process_age(df = helper_check_names(df))
}
\arguments{
\item{df}{A data frame containing an \code{age} column and other columns that start with "share". The \code{age} column can contain numeric values (0-120) or categorical values ("adult" or "kid").}
}
\value{
A tibble with three columns: \code{name}, \code{activity}, and \code{age}. The age values are transformed based on the rules provided, and the data is pivoted to a long format.
}
\description{
The \code{helper_process_age} function processes the \code{age} column in a given data frame. It performs assertive tests to validate the data, converts categorical age values ("adult" and "kid") into numeric equivalents, and scales numeric age values according to specified rules. Additionally, it pivots the data to transform share columns into long format and extracts activities.
}
\details{
The function performs the following steps:
\enumerate{
\item \strong{Assertive Tests}: Validates that the \code{age} column exists and that it contains only numeric values (0-120) or the categorical values "adult" and "kid".
\item \strong{Conversion}:
\itemize{
\item Converts "adult" to 1 and "kid" to 0.5.
\item Replaces \code{NA} and empty values with 1.
\item Scales numeric values above 1 by dividing by 18.
}
\item \strong{Data Transformation}: Pivots columns that start with "share" into a long format with \code{activity} and \code{age} columns.
}
}
\examples{
data("valid_trip_data", package = "costsplitter")
df <- valid_trip_data
data_age <- helper_process_age(df)

}
\seealso{
Other helper: 
\code{\link{helper_check_names}()},
\code{\link{helper_minimize_payments}()},
\code{\link{helper_process_adjustment}()},
\code{\link{helper_process_name}()},
\code{\link{helper_process_pay}()},
\code{\link{helper_process_share}()}
}
\concept{helper}
